AB#7728 Refactor FullPageSearch.tsx (#435)
